requirements.txt
fastapi>=0.122.0
uvicorn[standard]==0.30.6
starlette>=0.49.1
pydantic==2.12.4
pydantic-settings==2.12.0
slowapi==0.1.9
aiofiles==24.1.0
pymupdf==1.26.5
pdf2image==1.17.0
pytesseract==0.3.13
python-docx==1.1.2
python-pptx==1.0.2
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.20
numpy==2.2.6
opencv-python-headless==4.12.0.88
pyspellchecker==0.8.3

# pip>=25.3
